+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ Features
 Other
 -----
 * Cassandra 2.0 support removal (PYTHON-716)
+* cqlengine: disallow Counter create, save operations (PYTHON-497)
 
 3.12.0
 ======